Move test macros to common location and add processor test macro
