+++ src/processor.rs
@@ -74,69 +74,23 @@ pub enum ExecutionResult {
 
 #[cfg(test)]
 mod test {
+    use crate::test::macros::*;
     use crate::{csr::CSR32, instructions::Instruction, registers::Register};
     use pretty_assertions::assert_eq;
 
     use super::*;
 
-    /// Macro for creating a programme from a mixture of instructions and
-    /// pseudoinstructions.
-    macro_rules! instructions {
-        ($($instruction:expr),* $(,)*) => {{
-            let mut vec = Vec::new();
-            $(vec.extend($instruction);)*
-            vec
-        }};
-    }
-
-    /// Used by the macro above to make it possible to include instruction and
-    /// pseudoinstructions together as if of the same time in the vec type syntax
-    impl IntoIterator for Instruction {
-        type Item = Self;
-
-        type IntoIter = std::iter::Once<Self::Item>;
-
-        fn into_iter(self) -> Self::IntoIter {
-            std::iter::once(self)
-        }
-    }
-
     #[test]
-    // TODO: create macro for making mini programme test like this (reuse the
-    // existing test macros and move somewhere common.
     fn simple_run() {
-        // Arrange
-        let instructions = instructions![
-            Instruction::LI(Register::A0, 21),
-            Instruction::LI(Register::A1, 21),
+        processor_test!(
+            Instruction::LI(Register::A0, 21);
+            Instruction::LI(Register::A1, 21);
             Instruction::ADD {
                 rd: Register::A2,
                 rs1: Register::A1,
                 rs2: Register::A0,
             },
-        ];
-
-        let mut processor = Processor::<i32, CSR32>::default();
-        processor.store_instructions(0, instructions);
-
-        let initial_memory_state = processor.memory.clone();
-
-        // Act
-        processor.run::<Instruction>();
-
-        // Assert
-        let expected_final_state = Processor::<i32, CSR32> {
-            memory: initial_memory_state,
-            registers: Registers::<i32> {
-                a0: 21,
-                a1: 21,
-                a2: 42,
-                ..Default::default()
-            },
-            pc: 12,
-            ..Default::default()
-        };
-
-        assert_eq!(processor, expected_final_state)
+            results_in: {registers: {a0: 21, a1:21, a2: 42}, pc: 12},
+        );
     }
 }
