+++ src/test.rs
@@ -1,0 +1,1 @@
+pub(crate) mod macros;
